+++ src/sql/migrations.rs
@@ -54,6 +54,7 @@ mod tests {
         assert!(result.is_ok());
     }
 
+    #[derive(Debug, thiserror::Error)]
     pub enum Error {}
 
     pub struct TestAggregate;